requirements.txt
aiohttp~=3.7.4.post0
jsonschema~=3.2.0
python-dotenv~=0.19.0
pip~=21.1.3
cryptography~=3.4.8
setuptools~=56.0.0
pydantic~=1.8.2
attrs~=21.2.0
pdoc~=7.4.0
Jinja2~=3.0.1
Pygments~=2.10.0
MarkupSafe~=2.0.1
toml~=0.10.2
idna~=3.2
multidict~=5.1.0
six~=1.16.0
requests~=2.26.0
urllib3~=1.26.6
yarl~=1.6.3
isodate~=0.6.0
msrest~=0.6.21
isort~=5.9.3
colorama~=0.4.4
oauthlib~=3.1.1
pylint~=2.10.2
astroid~=2.7.3
platformdirs~=2.3.0
mccabe~=0.6.1
wrapt~=1.12.1
certifi~=2021.5.30
chardet~=4.0.0
pycparser~=2.20
pyrsistent~=0.18.0